Creating an Uninstall file.
Bro, if someone wants to re-install this software and its config, then first they need to remove this first. I just made the basic structure of the shell file. 
And after going through the installation process, I have mentioned the basic packages that it installs in the uninstall line. You could add more to it.
Thanks